# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Care Platform API",
        "description": "Complete API documentation for Care Platform - A comprehensive marketplace connecting service providers with customers",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Care Platform API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/reviews": {
            "get": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Get all reviews",
                "operationId": "1b9d177190f1105790093736fd25c348",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/reviews/pending": {
            "get": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Get pending reviews",
                "operationId": "ecae5fd2a7098d7a48ce2ca13509c954",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/reviews/flagged": {
            "get": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Get flagged reviews",
                "operationId": "bf96b00d0950a1d538774bb63c8dab15",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/reviews/{id}": {
            "get": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Get review details",
                "operationId": "65368a644a3796877e53335eb7327355",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Delete review",
                "operationId": "aa1553da346a90b8fe15bf8ed9e05a1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/reviews/{id}/approve": {
            "put": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Approve review",
                "operationId": "2200fd248af2922043914ea119dfdc1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approved successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/reviews/{id}/reject": {
            "put": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Reject review",
                "operationId": "e133b254f1782cd4b9863ce66a8e3a0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rejected successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/reviews/{id}/unflag": {
            "put": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Unflag review",
                "operationId": "5ff939d771eb25127d303c95b6cea9ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/reviews/statistics": {
            "get": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Get review statistics",
                "operationId": "1ca51a9f38ead0fd71d14452f9393126",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}